Split data users to user and users
